+++ Solar-Collector-Shape-Optimiser/solarcollector.hpp
@@ -7,9 +7,9 @@
 
 class SolarCollector : public Genome { // Inherits from Genome
 public:
-    ushort xsize;   // size of the panel
-    ushort ysize;
-    ushort hmax;  // maximal height (dictated by max printing height)
+    uint32_t xsize;   // size of the panel
+    uint32_t ysize;
+    uint32_t hmax;  // maximal height (dictated by max printing height)
 
     Mesh3d shape_mesh; // mesh calculated from 'shape' member
     vertex* mesh_midpoints; // precalculated mid points of every triangle of a mesh
@@ -19,22 +19,22 @@ public:
     Mesh3d* obstacle; // pointer to obstacle to read its mesh
 
 
-    SolarCollector (const unsigned short xs, const unsigned short ys, const unsigned short hm, Mesh3d* obs);
+    SolarCollector (const uint32_t xs, const uint32_t ys, const uint32_t hm, Mesh3d* obs);
     SolarCollector (const SolarCollector & other);
     SolarCollector (SolarCollector&& other) noexcept;
     SolarCollector& operator= (SolarCollector&& other) noexcept;
     SolarCollector& operator= (const SolarCollector & other);
     ~SolarCollector();
 
-    double getXY(const unsigned short x, const unsigned short y) const;
-    void setXY(const unsigned short x, const unsigned short y, const double val);
+    double getXY(const uint32_t x, const uint32_t y) const;
+    void setXY(const uint32_t x, const uint32_t y, const double val);
     void showYourself();
     
     bool rayIsBlocked(const vertex& source, const vertex& ray, const triangle& target);
     bool rayMeetsObstacle(const vertex& source, const vertex& ray, const triangle& target);
     void computeMesh();
     void computeMeshMidpoints();
-    void computeFitness(const vertex* rays, const int count_rays);
+    void computeFitness(const vertex* rays, const uint32_t count_rays);
     void exportAsSTL(std::string name);
     void exportReflectionAsSTL();
 };
